fix(orders): 테스트가 프로덕션 DB를 오염시키던 버그 수정 + 실제 주문내역 동기화 추가
## 근본 원인

tests/integration/test_toss_roundtrip.py의 test_signal_to_toss_order_roundtrip이
StateStore()를 인자 없이 생성했는데, 기본 경로가 프로덕션 DB
(~/quanteo/data/quanteo.db)를 가리킨다. MockTossRestClient(실제 Toss API를
호출하지 않는 가짜 클라이언트)로 executor.submit()을 호출하는 이 테스트가
pytest를 돌릴 때마다 가짜 'submitted' 주문을 실제 대시보드 주문내역에
그대로 삽입해왔다 (재현: 테스트 1회 실행 → orders 테이블 5행 → 6행).

## 수정

- test_signal_to_toss_order_roundtrip이 StateStore(":memory:")를 쓰도록 수정
- tests/conftest.py에 세션 전체 안전장치 추가 — 어떤 테스트든 프로덕션
  orders 테이블을 건드리면 즉시 실패시켜 같은 사고 재발을 원천 차단
- 기존에 쌓인 가짜 주문 6건은 DB에서 직접 정리 (백업 후 삭제)

## 추가 조사에서 드러난 구조적 문제

로컬 orders 테이블은 OrderExecutor.submit()이 만든 행만 반영하도록
설계됐는데, 이 submit()을 실제로 호출하는 신호 파이프라인이 없다
(StrategyEngine이 SIGNAL을 발행해도 RiskManager.evaluate()·
executor.submit()을 호출하는 구독자가 전혀 없음). Toss는 모의투자
구분이 없는 단일 계좌라 사용자가 Toss 앱에서 직접 낸 주문도 실제
계좌 상태의 일부인데, 로컬 DB만으로는 이를 반영할 수 없었다.

- core/execution/order_history_sync.py (신규): Toss list_orders()를
  브로커 source of truth로 삼아 orders 테이블 전체를 주기적으로
  동기화하는 OrderHistorySyncFeed 추가. 최초 1회 CLOSED 이력을 커서
  끝까지 페이지네이션해 백필하고, 이후엔 최신 페이지만 확인
  (ORDER_HISTORY Rate Limit 절약). --with-trading 여부와 무관하게
  항상 동작 (읽기 전용 GET만 수행)
- core/store/db.py: upsert_broker_order() 추가 — clientOrderId가 있으면
  기존 행 갱신, 없으면(Toss 앱에서 직접 낸 주문) toss-native-{orderId}로
  새로 삽입
- core/app.py: OrderHistorySyncFeed를 PositionSyncFeed와 같은 자리에 배선

## 백필 중 추가로 발견한 크래시

실제 운영 DB로 백필을 검증하다가 해외주식 fractional investing 체결
수량("0.000151")을 만나 _safe_int()가 RuntimeError로 죽는 걸 발견 —
positions.qty는 이미 REAL로 이 문제를 처리해뒀는데 orders.qty와
TossOrder.quantity/OrderExecution.filled_quantity는 int로 남아있던
누락이었다. 같은 패턴으로 float/REAL로 맞춰 수정 (회귀 테스트 포함).

실제 계정 데이터로 검증 결과 총 327건(cancelled 90·filled 211·
pending 3·rejected 23)의 실제 주문 이력이 정상적으로 반영됨을 확인.

+++ core/api/models.py
@@ -85,7 +85,7 @@ class OrderItem(BaseModel):
     env: str
     side: str  # "BUY" | "SELL"
     order_type: str  # "LIMIT" | "MARKET"
-    qty: int
+    qty: float  # 해외주식 fractional investing 지원 — 정수 아닐 수 있음
     price: Decimal
     status: str
     created_at: str
